refactor: unify feed editor view
